src/backend/requirements.txt
fastapi[standard]
uvicorn
python-dotenv
google-adk
soundfile
openai-whisper
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
fastmcp
asyncio
mcp
httpx
python-multipart